Refactor 405 Method Not Allowed handling
Because we need a custom header, we can either special-case 405 errors
in the error handling code or not use the error handler. This does the
latter because 405 errors can only arise early in the request processing
in a couple of places.

Also get rid of the restrictions in the API .cpp files to GET-only.
There's nothing wrong with a HEAD request against the nodes?nodes...
endpoint, and the rails port responds to these.

Overall the HEAD responses are now a lot closer to the rails port.

Fixes #37