Use array query for bbox in writeable backend.
See issue #68 for the background. Previously (in a direct port
from the Rails code), we used a complex system to build a "where"
clause like "tile = XXXX or tile between XXXX and YYYY" which was
difficult to read. As it turns out, a query like "tile =
ANY({XXXX})" seems to be equivalent in terms of performance.