Removed checked version of writeable bbox query.
This fixes issue #67.

Instead of checking whether the table is empty and using one of
two different queries, it now checks the assumption and uses the
unchecked query if it holds, throws an exception if it doesn't.

We shouldn't see the exception in real world usage, but it is
descriptive enough that someone experiencing it should be able to
file an issue about it.