Remove size_t num_elements from start_element_type
AMF is the only format that requires the formatter to know the
number of elements before serialization of those elements.

As we don't support AMF and don't plan to, there is no need for
this in the function call. This removes a bunch of sequential
scans of the tmp tables.

Fixes #74