Getting rid of unfinished AMF source file.
The plan was to support AMF/flash natively from cgimap, but it
seems the time has passed and flash is pretty much end-of-lined
by Adobe. Given the num_(nodes|ways|relations) calls have been
removed and the near zero chance of resurrecting this code, it's
better to not have it cluttering up the source tree.