Added static XML backend.
This is mainly for testing purposes, and loads an XML file into
RAM which it uses as a very inefficient database. This is OK for
testing and for small scale serving - perhaps demos. It's never
going to handle the full planet, but that's not what it's for.

I'm going to use this to mock the backend for the tests so that
most of them can avoid having to create new databases and slow
stuff like that.