Provide default implementations of `request` methods.
The `request` interface had methods to put data or flush the
output which were already implemented as part of the
`output_buffer` interface. These changes implement the put/flush
methods in terms of `output_buffer` and reduce the size of the
interface which needs to be implemented for subclasses.

Also, removed `accept_r` in the base interface, as this method is
only needed for FCGI requests and the derived type is known at
the point where it is used.