Changed interface to `request` to split out headers.
Instead of writing the status, headers and body manually to a
stream, which worked great for FCGI but isn't great for many other
request models, instead treat the status and headers specially.
This allows us to translate those calls into other request models,
for example; the Ruby on Rails request model.