Optionally enable C++11 mode, if the compiler supports it.
This is a sort-of fix for an issue with the memcached headers,
where they include `<cinttypes>`, which requires C++11 mode. Older
versions included `<tr1/cinttypes>`, which doesn't require C++11
mode. The "fix" here doesn't really fix it, but sanity suggests
that anything with memcached-requiring-C++11 would probably have a
C++11 compiler installed and therefore enabling that mode when it
is available should work around the problem.