Add a better libfcgi detection macro.
The new macro has better error messages, and doesn't assume
success as the last one did, so now emits an error if it can't
find either the fcgi headers or library. This should fix #100.

The new macro also doesn't assume a directory layout for
installed sources (it allows overriding both the library and
header locations), so is easier to use with installations where
the layout may be different.