Not Found results should be empty HTML documents.
The current behaviour of the Rails code is that missing elements cause an empty
HTML document to be returned to the client, regardless of whether or not the
X-Error-Format header has been set.

This is probably a combination of hysterical raisins, that the client for which
X-Error-Format was implemented was capable of receiving 404s, and that the
ActiveRecord::RecordNotFound error has no details about _which_ element was
missing (until Rails >=5.0.1).